chore: apply feedbacks (1 test failing)
