+++ test/zipObj.test.ts
@@ -11,6 +11,7 @@ expectType<{bool: boolean, number: number}>(zipObj(['bool', 'number'], [true, 1]
 expectType<{[key: string]: boolean}>(zipObj(['a'] as string[], [true, false] as boolean[]));
 expectType<{a: boolean}>(zipObj(['a'], [true, false] as boolean[]));
 expectType<{[key: string]: boolean}>(zipObj(['a'] as string[], [true, false]));
+expectType<{[key: string]: boolean | number | string}>(zipObj(['a', 'b', 'c'] as string[], [true, 1, 'a']));
 expectType<{bool: true, number: 1}>(zipObj(['bool', 'number'], [true, 1, 'string']));
 expectType<{bool: true, number: 1, missing: undefined}>(zipObj(['bool', 'number', 'missing'], [true, 1]));
